+++ src/components/Footer/Footer.css
@@ -6,7 +6,6 @@
   display: flex;
   justify-content: space-between;
   align-items: center;
-  /*  padding: 0 104px;*/
 }
 
 .footer__copyright {
@@ -55,7 +54,6 @@
 @media screen and (max-width: 800px) {
   .footer {
     min-height: 64px;
-    /*padding: 0 40px;*/
   }
 }
 
@@ -89,6 +87,6 @@
     margin-top: 38px;
     margin-bottom: 16px;
     width: 100%;
-    text-align: justify;
+    text-align: center;
   }
 }
