+++ src/components/MobileMenu/MobileMenu.css
@@ -20,7 +20,7 @@
   padding: 16px 16px 22px;
   max-height: 230px;
   width: 100%;
-  border-radius: 0 16px 16px;
+  border-radius: 0 0 16px 16px;
   background: #1a1b22;
 }
