+++ src/components/SavedNewsHeader/SavedNewsHeader.js
@@ -10,7 +10,7 @@ const SavedNewsHeader = () => {
     <section className='saved-news-header'>
       <h2 className='saved-news-header__title'>Сохранённые статьи</h2>
       <p className='saved-news-header__subtitle'>Грета, у вас 5 сохранённых статей</p>
-      <p className='saved-news-header__text'>По ключевым словам: <span className='saved-news-header__text-span'>Природа, Тайга и 2-м другим</span></p>
+      <p className='saved-news-header__text'>По ключевым словам: <span className='saved-news-header__text-span'>Природа, Тайга</span> и <span className='saved-news-header__text-span'>2-м другим</span></p>
     </section>
   );
